gx: improve GXPokeBlendMode store narrowing match

+++ src/gx/GXMisc.c
@@ -237,7 +237,7 @@ void GXPokeBlendMode(GXBlendMode type, GXBlendFactor src_factor, GXBlendFactor d
     reg = (reg & ~0xF000) | ((u32)op << 12);
     reg = (reg & ~0x0700) | ((u32)src_factor << 8);
     reg = (reg & ~0x00E0) | ((u32)dst_factor << 5);
-    *pe_reg_1 = reg;
+    *pe_reg_1 = (u16)reg;
 }
 
 void GXPokeColorUpdate(GXBool update_enable) {
